added evaluation of all metrics, moved __main__ to examples
