+++ python/plane_seg/metrics/__init__.py
@@ -1,2 +1,2 @@
-from . import evaluate_metrics
-from .evaluate_metrics import *
+from . import metrics_evaluation
+from .metrics_evaluation import *
