+++ python/plane_seg/metrics/examples/evaluate_all_metrics.py
@@ -1,0 +1,31 @@
+import sys
+from pathlib import Path
+
+from argparse import ArgumentParser
+import numpy as np
+
+from python.plane_seg.metrics import evaluate_all_metrics
+
+
+def main(argv):
+    parser = ArgumentParser()
+    parser.add_argument('--print_to_console', help='optional, true by default, can be one of: true, false')
+    parser.add_argument('--output_file', type=Path,
+                        help='optional, path to an output file.\nif specified, metric value is appended there')
+
+    args = parser.parse_args(argv)
+
+    if not args.predicted_labels or not args.ground_truth or not args.metric_name:
+        parser.print_help()
+        return
+
+    predicted_labels = np.load('data/labels.npy')
+
+    evaluate_all_metrics(predicted_labels,
+                         Path('data/0000.png'),
+                         print_to_console=(args.print_to_console != 'false'),
+                         output_file=args.output_file)
+
+
+if __name__ == '__main__':
+    main(sys.argv[1:])
